+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Unreleased
+**Features**:
+
+- Improve dynamic sampling rule configuration. ([#907](https://github.com/getsentry/relay/pull/907))
 
 **Internal**:
 
@@ -11,7 +14,6 @@
 **Features**:
 
 - Support dynamic sampling for error events. ([#883](https://github.com/getsentry/relay/pull/883))
-- Improve dynamic sampling rule configuration. ([#907](https://github.com/getsentry/relay/pull/907))
 
 **Bug Fixes**: